Add MockDataSeeder + bundled dog photos for portable local seeding
MockDataSeeder is standalone (NOT wired into DatabaseSeeder, so it never runs
on deploy) and populates a rich local dataset: ~58 dogs, ~74 users, 180
donations, plus applications, rescues, volunteers, visitations, reminders,
intakes and settings. It bundles dog1-9.jpg under database/seeders/dog-photos/
and copies them onto the public disk on run, so a fresh git clone can re-seed
with working images without transferring the gitignored storage/ folder.

Also fix create_app_notifications_table: user_id was a signed int (mirroring an
old imported DB) which broke a from-scratch `migrate` on MySQL (errno 150). It
is now unsignedBigInteger to match users.id; already-deployed tables are
unaffected and all 53 tests stay green.

+++ backend/database/migrations/2026_06_21_175910_create_app_notifications_table.php
@@ -15,10 +15,11 @@ return new class extends Migration
     {
         Schema::create('app_notifications', function (Blueprint $table) {
             $table->id();
-            // users.id is a signed `int(11)` on the live DB (not bigint unsigned), so the FK
-            // column must match that exact type or MySQL refuses to create the constraint
-            // (errno 150) — same gotcha noted in the volunteers migration.
-            $table->integer('user_id');
+            // Matches users.id (bigint unsigned, from $table->id()). This was previously a signed
+            // int to mirror an old imported DB whose users.id was int(11); on a clean `migrate`
+            // that mismatch breaks the FK (errno 150), so it now matches the real PK type. Safe on
+            // both MySQL and Postgres for fresh builds; existing deployed tables are unaffected.
+            $table->unsignedBigInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users')->cascadeOnDelete();
             $table->string('type', 50);
             $table->string('title', 150);
